requirements.txt
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
av==17.0.1
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.2.1
coloredlogs==15.0.1
ctranslate2==4.7.1
deprecation==2.1.0
dotenv==0.9.9
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
humanfriendly==10.0
idna==3.13
Jinja2==3.1.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lance-namespace==0.7.0
lance-namespace-urllib3-client==0.7.0
lancedb==0.25.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.20.0
networkx==3.6.1
numpy==1.26.4
onnxruntime==1.23.2
open_clip_torch==3.3.0
opencv-python-headless==4.9.0.80
overrides==7.7.0
packaging==26.1
pandas==3.0.2
pillow==12.1.1
platformdirs==4.9.6
protobuf==7.34.1
pyarrow==24.0.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
scenedetect==0.6.7.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
sympy==1.14.0
tdqm==0.0.1
tenacity==9.1.4
timm==1.0.26
tokenizers==0.22.2
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
